fix: 회원 탈퇴 시 비밀번호 검증 로직 추가
회원 탈퇴 요청 메서드 Patch -> Post로 변경.
비밀번호 검증 로직 추가 .

+++ src/main/java/com/example/newsfeed/domain/user/service/UserService.java
@@ -42,8 +42,11 @@ public class UserService {
 
     // 회원탈퇴
     @Transactional
-    public void withdrawal(Long id) {
+    public void withdrawal(Long id, String password) {
         User user = findByIdOrElseThrow(id);
+        if (!passwordEncoder.matches(password, user.getPassword())) {
+            throw new PasswordMismatchException("비밀번호가 올바르지 않습니다.");
+        }
         userRepository.delete(user);
     }
